requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
python-dateutil
APScheduler
httpx[http2]
pyjwt
cryptography

